Admin panel - order by

+++ activities/models.py
@@ -111,11 +111,11 @@ class Level(models.Model):
 class Venue(models.Model):
     
     name            = models.CharField(max_length=100, null=False)
-    street_nb       = models.CharField(max_length=20, null=False)
+    street_nb       = models.CharField(max_length=20,  null=False)
     street_name     = models.CharField(max_length=200, null=False)
     city            = models.CharField(max_length=100, null=False)
-    postcode        = models.PositiveSmallIntegerField(null=False)
-    iframe_url_map  = models.URLField(max_length=500, null=True)
+    postcode        = models.IntegerField(             null=False)
+    iframe_url_map  = models.URLField(max_length=500,  null=True)
     
     
     def __str__(self):
